split box\sphere collider component from base collider  component
